rename `src/contrib/beamnhl` --> `src/contrib/beamhnl`, add cpright
