--- src/contrib/beamnhl/CommonHNL_DEMO.xml
@@ -1,124 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-
-<!--
-******************************************************************************************************
-In this file, common physics parameters for the long-lived HNL simulation are defined.
-Ideally, the user would come here to tweak parameters such as HNL mass and mixings as well as switch
-on/off certain hooks such as Majorana nature, polarisation effects, etc.
-It would also be interesting to set the interesting decays here (relevant in case a specific channel is
-not specified in the arguments of the code itself, helps POT)
-
-This is a demo file meant for first-time use. You can use this by putting it in $GENIE/config/CommonHNL.xml and the fluxes and ROOT-geometries generated in $GENIE/src/contrib/beamhnl.
-It models a Dirac HNL of 200 MeV/c^2 mass and couplings |Ue4|^2 = |Um4|^2 = 10^-7, |U_t4|^2 = 0
-Signal channels are the two-body decays pi + mu, pi + e and the three-body invisible decay v + v + v.
-
-For simplicity, the detector frame USER is unrotated wrt the NEAR frame, placed 60 m downwards 1 km downstream of the target. The beam is rotated downwards by 0.05830 rad
-
-Author:
-John Plows <komninos-john.plows \at physics.ox.ac.uk>
-University of Oxford
-April 11th, 2022
-******************************************************************************************************
--->
-
-<common_HNL_list>
-
-  <param_set name="ParameterSpace">
-    <param type="double" name="HNL-Mass"> 0.200 </param> <!-- GeV -->
-    <param type="vec-double" name="HNL-LeptonMixing" delim=";"> 1.0e-7 ; 1.0e-7 ; 0.0 </param>
-    <param type="bool" name="HNL-Majorana"> false </param>
-    <param type="int" name="HNL-Type"> 2 </param> <!-- 0 = pure-N, 1 = pure-Nbar, 2 = mix -->
-    
-    <param type="bool" name="GetCMFrameInstead"> false </param>
-  </param_set>
-
-  <param_set name="InterestingChannels">
-    <!-- 2-body decays -->
-    <param type="bool" name="HNL-2B_mu_pi"> true </param>
-    <param type="bool" name="HNL-2B_e_pi"> true </param>
-    <param type="bool" name="HNL-2B_nu_pi0"> false </param>
-    <!-- 3-body decays -->
-    <param type="bool" name="HNL-3B_nu_nu_nu"> true </param>
-    <param type="bool" name="HNL-3B_nu_mu_mu"> false </param>
-    <param type="bool" name="HNL-3B_nu_e_e"> false </param>
-    <param type="bool" name="HNL-3B_nu_mu_e"> false </param>
-    <param type="bool" name="HNL-3B_e_pi_pi0"> false </param>
-    <param type="bool" name="HNL-3B_mu_pi_pi0"> false </param>
-    <param type="bool" name="HNL-3B_nu_pi0_pi0"> false </param>
-  </param_set>
-
-  <param_set name="InitialState">
-    <param type="double" name="HNL-angular_deviation"> 1.0 </param> <!-- deg -->
-    <param type="vec-double" name="HNL-energy_range" delim=";"> 0.0 ; 20.0 </param> <!-- GeV -->
-    <param type="double" name="HNL-max-energy"> 20.0 </param> <!-- GeV -->
-  </param_set>
-
-  <param_set name="CoordinateXForm">
-    <param type="vec-double" name="Near2Beam_R" delim=";"> 0.0 ; 0.0 ; -0.05830 </param> <!-- rad -->
-    <!-- Euler angles, extrinsic x-z-x = 1-2-3, RM * BEAM = USER, RM = Rx(1) * Rz(2) * Rx(3). -->
-    <!-- Describes rotation of BEAM wrt NEAR frame -->
-    <param type="vec-double" name="Near2User_T" delim=";"> 0.0 ; -60.0 ; 1000.0 </param> <!-- m -->
-    <!-- USER origin in NEAR coordinates -->
-    <param type="vec-double" name="Near2User_R" delim=";"> 0.0 ; 0.0 ; 0.0 </param>
-    <!-- Euler angles, extrinsic x-z-x -->
-    <!-- Describes rotation of USER wrt NEAR frame -->
-    <param type="vec-double" name="DetCentre_User" delim=";"> 0.0 ; 0.0 ; 0.0 </param> <!-- m -->
-    <!-- Position of detector centre in USER frame, in case it is not at USER origin -->
-  </param_set>
-
-  <param_set name="ParticleGun">
-    <param type="double" name="PG-OriginX"> 0.0 </param> <!-- user, m -->
-    <param type="double" name="PG-OriginY"> 0.0 </param> <!-- user, m -->
-    <param type="double" name="PG-OriginZ"> -10.0 </param> <!-- user, m -->
-
-    <param type="double" name="PG-OriginDX"> 1.0 </param> <!-- user, m, >= 0 -->
-    <param type="double" name="PG-OriginDY"> 1.0 </param> <!-- user, m, >= 0 -->
-    <param type="double" name="PG-OriginDZ"> 0.0 </param> <!-- user, m, >= 0 -->
-
-    <param type="double" name="PG-cx"> 0.0 </param> <!-- user, GeV/GeV -->
-    <param type="double" name="PG-cy"> 0.0 </param> <!-- user, GeV/GeV -->
-    <param type="double" name="PG-cz"> 1.0 </param> <!-- user, GeV/GeV -->
-
-    <param type="double" name="PG-DTheta"> 5.0 </param> <!-- deg, >= 0 -->
-    <param type="double" name="PG-DPhi"> 45.0 </param> <!-- deg, >= 0 -->
-
-    <param type="double" name="PG-Energy"> 1.0 </param> <!-- GeV -->
-  </param_set>
-
-  <param_set name="FluxCalc">
-    <param type="vec-double" name="ParentPOTScalings" delim=";"> 1.0 ; 1.005 ; 9.353 ; 10.458 </param> <!-- All parents -> no muons -> no pions -> no K0L -->
-    <param type="bool" name="IsParentOnAxis"> true </param>
-    <param type="bool" name="IncludePolarisation"> false </param>
-    <param type="bool" name="FixPolarisationDirection"> false </param>
-    <param type="vec-double" name="HNL-PolDir" delim=";"> 0.0 ; 0.0 ; 1.0 </param> 
-  </param_set>
-
-  <!--
-      <param_set name="Hooks">
-	  <param type="bool" name="UseRhos"> false </param> // for pi0+stuff in FS, use published decay width dominated by rho intermediate state. See the paper by Coloma et al.
-	  <param type="bool" name="CalcKaonFluxOnly"> false </param>
-	  <param type="bool" name="AttemptCharmCalculation"> false </param>
-	  <param type="bool" name="IncludeMultipleHNL"> false </param>
-      </param_set>
-  -->
-
-  <param_set name="External">
-    <!-- should these be migrated elsewhere? -->
-
-    <!-- from Coloma et al in EPJ C -->
-    <param type="double" name="Pion-FFactor"> 0.130 </param> <!-- GeV -->
-
-    <!-- www.nu-fit.org v5.1 w/ SK-atm data, no correlations -->
-    <!-- value = 1/2 * (min + max) -->
-    <param type="double" name="PMNS-Ue1"> 0.823 </param>
-    <param type="double" name="PMNS-Ue2"> 0.546 </param>
-    <param type="double" name="PMNS-Ue3"> 0.150 </param>
-    <param type="double" name="PMNS-Um1"> 0.3715 </param>
-    <param type="double" name="PMNS-Um2"> 0.599 </param>
-    <param type="double" name="PMNS-Um3"> 0.697 </param>
-    <param type="double" name="PMNS-Ut1"> 0.430 </param>
-    <param type="double" name="PMNS-Ut2"> 0.570 </param>
-    <param type="double" name="PMNS-Ut3"> 0.692 </param>
-  </param_set>
-
-</common_HNL_list>
